refactor: 사용되지 않는 클래스 제거
