+++ src/main/kotlin/omok/model/GameState.kt
@@ -1,7 +1,6 @@
 package omok.model
 
 enum class GameState {
-    WHITE_OMOK,
-    BLACK_OMOK,
     PLAYING,
+    FINISHED,
 }
