--- src/main/kotlin/omok/model/GameState2.kt
@@ -1,6 +1,0 @@
-package omok.model
-
-enum class GameState2 {
-    PLAYING,
-    FINISHED,
-}
