--- src/main/kotlin/omok/model/Points.kt
@@ -1,14 +1,0 @@
-package omok.model
-
-import rule.wrapper.point.Point
-
-class Points(
-    points: List<Point> = listOf(),
-) {
-    private var _points = points
-    val points get() = _points.map { it.copy() }
-
-    fun add(point: Point) {
-        _points += point
-    }
-}
