+++ src/main/kotlin/omok/model/Stone.kt
@@ -1,8 +1,6 @@
 package omok.model
 
-import rule.wrapper.point.Point
-
 data class Stone(
-    val point: Point,
+    val position: Position,
     val color: Color,
 )
