--- src/main/kotlin/omok/model/Stone2.kt
@@ -1,6 +1,0 @@
-package omok.model
-
-data class Stone2(
-    val position: Position,
-    val color: Color,
-)
