--- src/main/kotlin/omok/model/WhitePlayer.kt
@@ -1,21 +1,0 @@
-package omok.model
-
-import rule.OmokRule
-import rule.WhiteRenjuRule
-import rule.wrapper.point.Point
-
-class WhitePlayer : Player() {
-    override val points = Points()
-    override val rule: OmokRule = WhiteRenjuRule()
-
-    override fun checkGameState(newPoint: Point): GameState {
-        return when (rule.checkSerialSameStonesBiDirection(points.points, newPoint, OMOK_CONDITION)) {
-            true -> GameState.WHITE_OMOK
-            false -> GameState.PLAYING
-        }
-    }
-
-    companion object {
-        private const val OMOK_CONDITION = 5
-    }
-}
